+++ lib/Producer.d.ts
@@ -19,6 +19,12 @@ export interface ProducerOptions {
      */
     paused?: boolean;
     /**
+     * Just for video. Time (in ms) before asking the sender for a new key frame
+     * after having asked a previous one. Default 1000. Minimum valid value is
+     * 500.
+     */
+    keyFrameWaitTime?: number;
+    /**
      * Custom application data.
      */
     appData?: any;