+++ test/test-Producer.js
@@ -199,7 +199,8 @@ test('transport2.produce() succeeds', async () =>
 					cname : 'video-1'
 				}
 			},
-			appData : { foo: 1, bar: '2' }
+			keyFrameWaitTime : 500,
+			appData          : { foo: 1, bar: '2' }
 		});
 
 	expect(onObserverNewProducer).toHaveBeenCalledTimes(1);
@@ -334,6 +335,33 @@ test('transport1.produce() with wrong arguments rejects with TypeError', async (
 					cname : 'video-1'
 				}
 			}
+		}))
+		.rejects
+		.toThrow(TypeError);
+
+	// Wrong keyFrameWaitTime (less than 500).
+	await expect(transport1.produce(
+		{
+			kind          : 'video',
+			rtpParameters :
+			{
+				codecs :
+				[
+					{
+						mimeType    : 'video/h264',
+						payloadType : 112,
+						clockRate   : 90000,
+						parameters  :
+						{
+							'packetization-mode' : 1,
+							'profile-level-id'   : '4d0032'
+						}
+					}
+				],
+				encodings : [ { ssrc: 6666 } ],
+				rtcp      : { cname: 'qwerty' }
+			},
+			keyFrameWaitTime : 499
 		}))
 		.rejects
 		.toThrow(TypeError);