+++ worker/src/RTC/Producer.cpp
@@ -258,7 +258,27 @@ namespace RTC
 
 		// Create a KeyFrameRequestManager.
 		if (this->kind == RTC::Media::Kind::VIDEO)
-			this->keyFrameRequestManager = new RTC::KeyFrameRequestManager(this);
+		{
+			auto jsonKeyFrameWaitTimeIt = data.find("keyFrameWaitTime");
+			uint32_t keyFrameWaitTime   = 1000;
+
+			// clang-format off
+			if (
+				jsonKeyFrameWaitTimeIt != data.end() &&
+				jsonKeyFrameWaitTimeIt->is_number_integer()
+			)
+			// clang-format on
+			{
+				keyFrameWaitTime = jsonKeyFrameWaitTimeIt->get<uint32_t>();
+			}
+
+			if (keyFrameWaitTime < 500)
+			{
+				MS_THROW_TYPE_ERROR("keyFrameWaitTime must be higher or equal than 500 ms");
+			}
+
+			this->keyFrameRequestManager = new RTC::KeyFrameRequestManager(this, keyFrameWaitTime);
+		}
 	}
 
 	Producer::~Producer()