+++ worker/test/src/RTC/TestKeyFrameRequestManager.cpp
@@ -33,7 +33,7 @@ SCENARIO("KeyFrameRequestManager", "[rtp][keyframe]")
 	SECTION("key frame requested once, not received on time")
 	{
 		listener.Reset();
-		KeyFrameRequestManager keyFrameRequestManager(&listener);
+		KeyFrameRequestManager keyFrameRequestManager(&listener, 1000);
 
 		keyFrameRequestManager.KeyFrameNeeded(1111);
 
@@ -45,7 +45,7 @@ SCENARIO("KeyFrameRequestManager", "[rtp][keyframe]")
 	SECTION("key frame requested many times, not received on time")
 	{
 		listener.Reset();
-		KeyFrameRequestManager keyFrameRequestManager(&listener);
+		KeyFrameRequestManager keyFrameRequestManager(&listener, 500);
 
 		keyFrameRequestManager.KeyFrameNeeded(1111);
 		keyFrameRequestManager.KeyFrameNeeded(1111);
@@ -60,7 +60,7 @@ SCENARIO("KeyFrameRequestManager", "[rtp][keyframe]")
 	SECTION("key frame is received on time")
 	{
 		listener.Reset();
-		KeyFrameRequestManager keyFrameRequestManager(&listener);
+		KeyFrameRequestManager keyFrameRequestManager(&listener, 500);
 
 		keyFrameRequestManager.KeyFrameNeeded(1111);
 		keyFrameRequestManager.KeyFrameReceived(1111);
@@ -73,7 +73,7 @@ SCENARIO("KeyFrameRequestManager", "[rtp][keyframe]")
 	SECTION("key frame is forced, no received on time")
 	{
 		listener.Reset();
-		KeyFrameRequestManager keyFrameRequestManager(&listener);
+		KeyFrameRequestManager keyFrameRequestManager(&listener, 500);
 
 		keyFrameRequestManager.KeyFrameNeeded(1111);
 		keyFrameRequestManager.ForceKeyFrameNeeded(1111);
@@ -86,7 +86,7 @@ SCENARIO("KeyFrameRequestManager", "[rtp][keyframe]")
 	SECTION("key frame is forced, received on time")
 	{
 		listener.Reset();
-		KeyFrameRequestManager keyFrameRequestManager(&listener);
+		KeyFrameRequestManager keyFrameRequestManager(&listener, 500);
 
 		keyFrameRequestManager.KeyFrameNeeded(1111);
 		keyFrameRequestManager.ForceKeyFrameNeeded(1111);